Lab8/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-api-python-client==2.97.0
python-dotenv==1.0.0
jinja2==3.1.2
pydantic==2.3.0
google-generativeai>=0.3.0
groq>=0.4.0